Adjust label template
If the label is for a container then if the label is set to be visible render it as a span, otherwise don't render it at all. This is better semantically for containers, which are divs, and prevents screen readers from reading out hidden labels.

n.b. Building this code also causes an update to the .govuk-form-input--error class. This uses the $govuk-border-width-form-element-error variable for border width which was reduced from 4px to 2px in govuk-frontend in July 2020.